refactor: 페이지 단위 컴포넌트 pages폴더로 분리 및 네이밍 변경
- CardPaymentsPage -> CardForm
