+++ src/main.tsx
@@ -1,9 +1,9 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import CardPaymentsPage from "./components/Card/CardPaymentsPage/CardPaymentsPage";
+import CardForm from "./pages/CardForm/CardForm";
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
   <React.StrictMode>
-    <CardPaymentsPage />
+    <CardForm />
   </React.StrictMode>
 );
